requirements.txt
Flask==2.2.2
protobuf==4.23.4
python-dotenv==1.0.1
Requests==2.31.0
Werkzeug==2.2.3
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.1.17
langchain-google-genai==0.0.6
langsmith==0.0.87
google-generativeai==0.4.0